Move QueryGridModel to separate file.
- Also move some tests to more appropriate files
