+++ packages/components/src/internal/components/editable/EditableGrid.tsx
@@ -38,8 +38,7 @@ import { headerSelectionCell } from '../../renderers';
 import { QueryInfoForm, QueryInfoFormProps } from '../forms/QueryInfoForm';
 import { MAX_EDITABLE_GRID_ROWS } from '../../constants';
 import { Grid, GridColumn } from '../base/Grid';
-import { GRID_CHECKBOX_OPTIONS, GRID_EDIT_INDEX, GRID_SELECTION_INDEX } from '../base/models/constants';
-import { QueryGridModel } from '../base/models/model';
+import { GRID_CHECKBOX_OPTIONS, GRID_EDIT_INDEX, GRID_SELECTION_INDEX } from '../../constants';
 import { blurActiveElement, capitalizeFirstChar, caseInsensitive } from '../../util/utils';
 import { DeleteIcon } from '../base/DeleteIcon';
 import { Alert } from '../base/Alert';
@@ -47,7 +46,7 @@ import { LoadingSpinner } from '../base/LoadingSpinner';
 import { EditorModel, ValueDescriptor } from '../../models';
 import { BulkAddUpdateForm } from '../forms/BulkAddUpdateForm';
 
-import { QueryColumn } from '../../..';
+import { QueryColumn, QueryGridModel } from '../../..';
 
 import { AddRowsControl, AddRowsControlProps, PlacementType } from './Controls';
 import { Cell } from './Cell';
