+++ packages/components/src/internal/components/editable/EditableGridLoader.tsx
@@ -16,9 +16,8 @@
 import React from 'react';
 import { fromJS, List } from 'immutable';
 
+import { IGridLoader, IGridResponse, selectRows, QueryGridModel } from '../../..';
 import { EditorModel } from '../../models';
-import { selectRows } from '../../query/api';
-import { IGridLoader, IGridResponse, QueryGridModel } from '../base/models/model';
 
 export class EditableGridLoader implements IGridLoader {
     fetch(gridModel: QueryGridModel): Promise<IGridResponse> {
