+++ packages/components/src/internal/components/menus/SelectionMenuItem.tsx
@@ -16,8 +16,7 @@
 import React, { PureComponent } from 'react';
 import { MenuItem, OverlayTrigger, Popover } from 'react-bootstrap';
 
-import { QueryGridModel } from '../base/models/model';
-import { QueryModel } from '../../..';
+import { QueryModel, QueryGridModel } from '../../..';
 
 interface Props {
     id: string;
