+++ packages/components/src/public/QueryModel/GridPanel.tsx
@@ -13,8 +13,9 @@ import {
     QueryConfig,
     QueryInfo,
     QuerySort,
+    Pagination,
 } from '../..';
-import { GRID_SELECTION_INDEX } from '../../internal/components/base/models/constants';
+import { GRID_SELECTION_INDEX } from '../../internal/constants';
 import { headerCell, headerSelectionCell } from '../../internal/renderers';
 import { ActionValue } from '../../internal/components/omnibox/actions/Action';
 import { FilterAction } from '../../internal/components/omnibox/actions/Filter';
@@ -22,7 +23,6 @@ import { SearchAction } from '../../internal/components/omnibox/actions/Search';
 import { SortAction } from '../../internal/components/omnibox/actions/Sort';
 import { ViewAction } from '../../internal/components/omnibox/actions/View';
 import { Change, ChangeType, OmniBox } from '../../internal/components/omnibox/OmniBox';
-import { Pagination } from '../../internal/components/pagination/Pagination';
 
 import { InjectedQueryModels, RequiresModelAndActions, withQueryModels } from './withQueryModels';
 import { ViewMenu } from './ViewMenu';
