+++ packages/components/src/public/QueryModel/QueryModel.ts
@@ -12,8 +12,8 @@ import {
     SchemaQuery,
     ViewInfo,
 } from '../..';
-import { GRID_SELECTION_INDEX } from '../../internal/components/base/models/constants';
-import { PaginationData } from '../../internal/components/pagination/Pagination';
+import { GRID_SELECTION_INDEX } from '../../internal/constants';
+import { PaginationData } from '../..';
 import { DataViewInfo } from '../../internal/models';
 
 import { flattenValuesFromRow, offsetFromString, querySortsFromString, searchFiltersFromString } from './utils';
