+++ packages/components/src/typings/global.d.ts
@@ -18,8 +18,7 @@ import { List, Map } from 'immutable';
 import { User } from '@labkey/api';
 
 import { LookupStore, EditorModel } from '../internal/models';
-import { QueryGridModel } from '../internal/components/base/models/model';
-import { NotificationItemModel } from '../internal/components/notifications/model';
+import { NotificationItemModel, QueryGridModel } from '..';
 
 // Typescript helper definition of global state for reactn
 declare module 'reactn/default' {
